Pull

+++ js/saved.js
@@ -1,18 +1,11 @@
 /// <reference path="common.js" />
-/* global modalPrompt, setupOfflineIndicator, modalConfirm, getModalInputText,
-   isEmbedded, isIOSEdge, exportSvg, exportPng, StoreName, isIOS, setupModal */
+/* global showModalDialog, getModalInputText, isEmbedded, exportSvg, exportPng,
+   StoreName, isIOS, setupModal */
 
 /** Creates a visual divider for the item list. */
 function newDivider() {
   const element = document.createElement('div');
   element.className = 'divider';
-  return element;
-}
-
-/** Creates a container div to put the action buttons in. */
-function newActions() {
-  const element = document.createElement('div');
-  element.className = 'actions';
   return element;
 }
 
@@ -28,6 +21,7 @@ function newDiagramLink(text, href) {
 /** Creates an 'Export PNG' button. */
 function newPngButton() {
   const pngLink = document.createElement('button');
+  pngLink.type = 'button';
   pngLink.className = 'export';
   pngLink.title = 'Export PNG';
   pngLink.textContent = 'PNG';
@@ -37,6 +31,7 @@ function newPngButton() {
 /** Creates an 'Export SVG' button. */
 function newSvgButton() {
   const svgLink = document.createElement('button');
+  svgLink.type = 'button';
   svgLink.className = 'export';
   svgLink.title = 'Export SVG';
   svgLink.textContent = 'SVG';
@@ -46,6 +41,7 @@ function newSvgButton() {
 /** Creates a 'Delete' button. */
 function newDeleteButton() {
   const deleteImage = document.createElement('button');
+  deleteImage.type = 'button';
   deleteImage.className = 'delete';
   deleteImage.title = 'Delete';
   return deleteImage;
@@ -54,6 +50,7 @@ function newDeleteButton() {
 /** Creates a 'Rename' button. */
 function newRenameButton() {
   const renameImage = document.createElement('button');
+  renameImage.type = 'button';
   renameImage.className = 'rename';
   renameImage.title = 'Rename';
   return renameImage;
@@ -68,39 +65,50 @@ function addEmptyMessage(container) {
 
 /** Adds a new row to the saved diagrams list. */
 function addNewRow(container, key, entry) {
-  const entryLink = newDiagramLink(entry.Title, '/viewsvg.htm#*' + key);
+  const entryLink = newDiagramLink(entry.Title, `/viewsvg.htm#*${key}`);
   container.appendChild(entryLink);
 
-  const entryActions = newActions();
-  const thisDivider = newDivider();
-
   const entryRename = newRenameButton();
-  entryActions.appendChild(entryRename);
+  container.appendChild(entryRename);
+
+  const entryDelete = newDeleteButton();
+  container.appendChild(entryDelete);
+
+  const entrySvgButton = newSvgButton();
+  container.appendChild(entrySvgButton);
+
+  const entryPngButton = newPngButton();
+  container.appendChild(entryPngButton);
+
+  const thisDivider = newDivider();
+  container.appendChild(thisDivider);
+
   entryRename.addEventListener('click', function renameClick() {
-    modalPrompt('Rename saved diagram:', entryLink.textContent,
-      function renameOK() {
+    showModalDialog('Rename saved diagram:', true, entryLink.textContent,
+      'OK', function renameOK() {
         const newName = getModalInputText();
         if (!newName) return;
-
         entry.Title = newName;
         const entryJSON = JSON.stringify(entry);
         localStorage.setItem(key, entryJSON);
         entryLink.textContent = newName;
-      });
+      },
+      undefined, undefined,
+      'Cancel', undefined);
   });
 
-  const entryDelete = newDeleteButton();
-  entryActions.appendChild(entryDelete);
   entryDelete.addEventListener('click', function deleteClick() {
-    const message =
-      `Are you sure you want to delete "${entryLink.textContent}"?`;
-
-    modalConfirm(message,
-      function deleteYes() {
+    showModalDialog(
+      `Are you sure you want to delete "${entryLink.textContent}"?`,
+      false, undefined,
+      'Yes', function deleteYes() {
         localStorage.removeItem(key);
 
         container.removeChild(entryLink);
-        container.removeChild(entryActions);
+        container.removeChild(entryRename);
+        container.removeChild(entryDelete);
+        container.removeChild(entrySvgButton);
+        container.removeChild(entryPngButton);
         container.removeChild(thisDivider);
 
         if (container.getElementsByClassName('diagram').length === 0) {
@@ -109,32 +117,25 @@ function addNewRow(container, key, entry) {
           addEmptyMessage(container);
           container.appendChild(newDivider());
         }
-      });
+      },
+      undefined, undefined,
+      'No', undefined);
   });
 
-  if (!isIOSEdge) {
-    const entrySvgButton = newSvgButton();
-    entryActions.appendChild(entrySvgButton);
-    entrySvgButton.addEventListener('click',
-      function svgLinkClick(event) {
-        event.preventDefault();
-        exportSvg(entry.Title + '.svg', entry.SvgXml);
-      }
-    );
+  entrySvgButton.addEventListener('click',
+    function svgLinkClick(event) {
+      event.preventDefault();
+      exportSvg(entry.Title + '.svg', entry.SvgXml);
+    }
+  );
 
-    const entryPngButton = newPngButton();
-    entryActions.appendChild(entryPngButton);
-    entryPngButton.addEventListener('click',
-      function pngLinkClick() {
-        const background = window.getComputedStyle(document.body)
-          .backgroundColor;
-        exportPng(entry.Title + '.png', entry.SvgXml, background);
-      }
-    );
-  }
-
-  container.appendChild(entryActions);
-  container.appendChild(thisDivider);
+  entryPngButton.addEventListener('click',
+    function pngLinkClick() {
+      const background =
+        window.getComputedStyle(document.body).backgroundColor;
+      exportPng(entry.Title + '.png', entry.SvgXml, background);
+    }
+  );
 }
 
 /** Adds all the saved diagrams onto the page. */
@@ -147,7 +148,9 @@ function populateList() {
   let keys = [];
   for (let index = 0; index < localStorage.length; index += 1) {
     const key = localStorage.key(index);
-    if (key !== StoreName.Flags && key !== StoreName.Settings) {
+    if (key !== StoreName.Flags &&
+        key !== StoreName.Settings &&
+        key !== StoreName.MatrixSelection) {
       keys.push(key);
     }
   }
@@ -158,9 +161,7 @@ function populateList() {
     }
   );
 
-  for (let index = 0; index < keys.length; index += 1) {
-    const key = keys[index];
-
+  for (const key of keys) {
     const entryJSON = localStorage.getItem(key);
     if (entryJSON) {
       const entry = JSON.parse(entryJSON);
@@ -176,48 +177,61 @@ function populateList() {
   }
 }
 
+/** Imports the file at index from the files list. */
+function processImport(files, index) {
+  const file = files[index];
+
+  const reader = new FileReader();
+  reader.addEventListener('load',
+    function fileLoaded(event) {
+      showModalDialog(
+        files.length === 0 ?
+          'Import diagram as:' :
+          `Import diagram ${index + 1} of ${files.length} as:`,
+        true, file.name,
+        'OK', function importOK() {
+          const diagramTitle = getModalInputText();
+          if (!diagramTitle) return;
+
+          const storageKey = Date.now().toString();
+          const svgObject = {
+            Title: diagramTitle,
+            SvgXml: event.target.result,
+          };
+
+          const jsonData = JSON.stringify(svgObject);
+          localStorage.setItem(storageKey, jsonData);
+
+          populateList();
+
+          if (index !== files.length - 1) {
+            index++;
+            processImport(files, index);
+          }
+        },
+        undefined, undefined,
+        files.length === 1 ? 'Cancel' : 'Skip',
+        function importSkipOrCancel() {
+          if (index !== files.length - 1) {
+            processImport(files, ++index);
+          }
+        });
+    }
+  );
+
+  reader.readAsText(file);
+}
+
 /** Imports the selected files into the Saved Diagrams list. */
 function filesSelected(event) {
-  const reader = new FileReader();
-
-  const fileList = event.target.files;
-  for (let i = 0; i < fileList.length; i++) {
-    const file = fileList[i];
-
-    reader.addEventListener('load',
-      function fileLoaded(event) {
-        modalPrompt('Import diagram as:', file.name,
-          function importOK() {
-            const diagramTitle = getModalInputText();
-            if (!diagramTitle) return;
-
-            const storageKey = Date.now().toString();
-            const svgObject = {
-              Title: diagramTitle,
-              SvgXml: event.target.result
-            };
-
-            const jsonData = JSON.stringify(svgObject);
-            localStorage.setItem(storageKey, jsonData);
-
-            const container = document.getElementById('collection');
-            addNewRow(container, storageKey, svgObject);
-          });
-      }
-    );
-
-    reader.readAsText(file);
-  }
-
-  event.target.value = '';
+  processImport(event.target.files, 0);
 }
 
 /** Clicks the export link corresponding with the supplied exportType for
  * all saved diagrams. */
 function exportAllDiagrams(exportType) {
   const exportLinks = document.getElementsByClassName('export');
-  for (let i = 0; i < exportLinks.length; i++) {
-    const exportLink = exportLinks[i];
+  for (const exportLink of exportLinks) {
     if (exportLink.textContent === exportType) {
       exportLink.click();
     }
@@ -230,29 +244,28 @@ function DOMContentLoaded() {
     document.getElementById('menu').style.display = 'none';
   }
 
-  setupOfflineIndicator();
   setupModal();
   populateList();
 
   if (isIOS) {
     document.getElementById('export-all-png').style.display = 'none';
   } else {
-    document.getElementById('export-all-png').addEventListener('click', () =>
-      exportAllDiagrams('PNG'));
+    document.getElementById('export-all-png').addEventListener('click',
+      () => exportAllDiagrams('PNG'));
   }
 
   if (isIOS) {
     document.getElementById('export-all-svg').style.display = 'none';
   } else {
-    document.getElementById('export-all-svg').addEventListener('click', () =>
-      exportAllDiagrams('SVG'));
+    document.getElementById('export-all-svg').addEventListener('click',
+      () => exportAllDiagrams('SVG'));
   }
 
-  document.getElementById('import').addEventListener('click', () =>
-    document.getElementById('file-selector').click());
+  document.getElementById('import').addEventListener('click',
+    () => document.getElementById('file-selector').click());
 
-  document.getElementById('file-selector')
-    .addEventListener('change', filesSelected);
+  document.getElementById('file-selector').
+    addEventListener('change', filesSelected);
 }
 
 document.addEventListener('DOMContentLoaded', DOMContentLoaded);
